Feat: Support InfluxDB 2.x
