+++ MyPerf4J-Base/src/main/java/cn/myperf4j/base/influxdb/InfluxDbClientFactory.java
@@ -10,9 +10,20 @@ public final class InfluxDbClientFactory {
 
     private static final InfluxDbConfig CONFIG = ProfilingConfig.influxDBConfig();
 
-    private static final InfluxDbClient CLIENT = new InfluxDbClient.Builder()
+    private static final InfluxDbV1Client V1_CLIENT = new InfluxDbV1Client.Builder()
             .host(CONFIG.host())
             .port(CONFIG.port())
+            .database(CONFIG.database())
+            .username(CONFIG.username())
+            .password(CONFIG.password())
+            .connectTimeout(CONFIG.connectTimeout())
+            .readTimeout(CONFIG.readTimeout())
+            .build();
+
+    private static final InfluxDbV2Client V2_CLIENT = new InfluxDbV2Client.Builder()
+            .host(CONFIG.host())
+            .port(CONFIG.port())
+            .organized(CONFIG.organized())
             .database(CONFIG.database())
             .username(CONFIG.username())
             .password(CONFIG.password())
@@ -25,6 +36,14 @@ public final class InfluxDbClientFactory {
     }
 
     public static InfluxDbClient getClient() {
-        return CLIENT;
+        return CONFIG.version().startsWith("2.") ? getV2Client() : getV1Client();
+    }
+
+    public static InfluxDbV1Client getV1Client() {
+        return V1_CLIENT;
+    }
+
+    public static InfluxDbV2Client getV2Client() {
+        return V2_CLIENT;
     }
 }
