+++ MyPerf4J-Base/src/test/java/cn/myperf4j/base/influxdb/InfluxDbV1ClientTest.java
@@ -7,9 +7,9 @@ import java.util.concurrent.TimeUnit;
 /**
  * Created by LinShunkang on 2020/05/19
  */
-public class InfluxDbClientTest {
+public class InfluxDbV1ClientTest {
 
-    private final InfluxDbClient influxDbClient = new InfluxDbClient.Builder()
+    private final InfluxDbV1Client influxDbV1Client = new InfluxDbV1Client.Builder()
             .host("127.0.0.1")
             .port(8086)
             .connectTimeout(100)
@@ -21,13 +21,13 @@ public class InfluxDbClientTest {
 
     @Test
     public void testBuildDatabase() {
-        boolean database = influxDbClient.createDatabase();
+        boolean database = influxDbV1Client.createDatabase();
         System.out.println(database);
     }
 
     @Test
     public void testWrite() throws InterruptedException {
-        boolean write = influxDbClient.writeMetricsAsync(
+        boolean write = influxDbV1Client.writeMetricsAsync(
                 "cpu_load_short,host=server01,region=us-west value=0.64 1434055562000000000\n" +
                         "cpu_load_short,host=server02,region=us-west value=0.96 1434055562000000000");
         System.out.println(write);
